+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Yii3 Voyti 2FA Changelog
 
-## 1.0.2 under development
+## 1.0.2 - August 30, 2026
 
 - New: Add `voyti:2fa:disable` console command to disable two-factor authentication for a user.
 - Bug: Avoid double-dispatching `BeforeLoginEvent` by calling `LoginCompletionService::finalize()` instead of `::complete()` when completing a 2FA-confirmed login.